+++ lib/src/cli/commands/export/export.dart
@@ -1,0 +1,1 @@
+export 'export_spicetify_command.dart';