+++ lib/src/cli/runner.dart
@@ -25,6 +25,7 @@ CommandRunner<int> createInPhaseCommandRunner({
         ..addCommand(CrawlCommand())
         ..addCommand(CurateCommand())
         ..addCommand(CuesCommand())
+        ..addCommand(ExportCommand())
         ..addCommand(LibraryCommand())
         ..addCommand(LoginCommand())
         ..addCommand(SearchCommand())